component-entrypoints
- `stepper`: добавлена точка входа `index.ts` (patch)
- `stroked-svg`: добавлена точка входа `index.ts` (patch)
- `super-ellipse-clip-path`: добавлена точка входа `index.ts` (patch)
- `switcher-row`: добавлена точка входа `index.ts` (patch)
- `tabs`: добавлена точка входа `index.ts` (patch)
