+++ src/stepper/index.ts
@@ -1,0 +1,2 @@
+export type { StepperSize, StepperStyle, StepperProps } from './types';
+export { Stepper } from './stepper';
