+++ src/stroked-svg/index.ts
@@ -1,0 +1,2 @@
+export type { StrokedSVGProps } from './types';
+export { StrokedSVG } from './stroked-svg';
