+++ src/stroked-svg/stroked-svg.module.scss
