+++ src/super-ellipse-clip-path/index.ts
@@ -1,0 +1,2 @@
+export type { SuperEllipseClipPathProps } from './types';
+export { SuperEllipseClipPath } from './super-ellipse-clip-path';
