+++ src/super-ellipse-clip-path/types.ts
@@ -1,0 +1,4 @@
+export interface SuperEllipseClipPathProps {
+  /** Идентификатор. */
+  id?: string;
+}
