+++ src/switcher-row/index.ts
@@ -1,0 +1,2 @@
+export type { SwitcherRowProps } from './types';
+export { SwitcherRow } from './switcher-row';
