+++ src/tabs/index.ts
@@ -1,0 +1,2 @@
+export type { TabsStyle, TabsProps, TabProps } from './types';
+export { Tabs, Tab } from './tabs';
